+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 <!--next-version-placeholder-->
 
+## v3.4.0 (2022-12-02)
+### Feature
+* Add auto seat climate command and properties ([#371](https://github.com/zabuldon/teslajsonpy/issues/371)) ([`21c5ff3`](https://github.com/zabuldon/teslajsonpy/commit/21c5ff3306d3caee48fa5062bb6194be1471f51a))
+
+## v3.3.0 (2022-11-30)
+### Feature
+* Add remote start command ([#367](https://github.com/zabuldon/teslajsonpy/issues/367)) ([`350f48b`](https://github.com/zabuldon/teslajsonpy/commit/350f48b47fbd885c452081a1ecf6ecc71b491e11))
+* Add valet mode command and property ([#368](https://github.com/zabuldon/teslajsonpy/issues/368)) ([`5e7c69f`](https://github.com/zabuldon/teslajsonpy/commit/5e7c69f848515b0d72362daaac28824081751878))
+
+## v3.2.2 (2022-11-22)
+### Fix
+* Handle unavailable sites/vehicles ([#365](https://github.com/zabuldon/teslajsonpy/issues/365)) ([`2daef7b`](https://github.com/zabuldon/teslajsonpy/commit/2daef7b3a7fbdbceb082749abec2386f6d97d556))
+
+## v3.2.1 (2022-11-21)
+### Fix
+* Keep existing vehicle config when asleep/offline ([#363](https://github.com/zabuldon/teslajsonpy/issues/363)) ([`92774e9`](https://github.com/zabuldon/teslajsonpy/commit/92774e975d9a5f3438d30c330f9e31481fc7b915))
+
+## v3.2.0 (2022-11-20)
+### Feature
+* Add usable_battery_level method in TeslaCar #361 ([`abf0d08`](https://github.com/zabuldon/teslajsonpy/commit/abf0d085ac1b6b26b85f081b838dc72abfe616b2))
+
 ## v3.1.0 (2022-11-02)
 ### Feature
 * Add window control and binary sensor ([#357](https://github.com/zabuldon/teslajsonpy/issues/357)) ([`29fabf7`](https://github.com/zabuldon/teslajsonpy/commit/29fabf726bc589dacf2c3fc6cd8432a776eaf6d9))